% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guam_BB_expansion.R
\name{calc_var_average}
\alias{calc_var_average}
\title{Variance of the mean of a variable}
\usage{
calc_var_average(x)
}
\arguments{
\item{x}{A vector of values}
}
\value{
Variance of the average of those values
}
\description{
Variance of the mean of a variable
}
