% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guam_BB_expansion.R
\name{calc_var_product}
\alias{calc_var_product}
\title{Variance of product x and y}
\usage{
calc_var_product(x, var_x, y, var_y)
}
\arguments{
\item{x}{Point estimate of x}

\item{var_x}{Variance of x}

\item{y}{Point estimate of y}

\item{var_y}{Variance of y}
}
\value{
the variance of the product of the two variables
}
\description{
Variance of product x and y
}
