% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guam_BB_expansion.R
\name{calc_var_quotient}
\alias{calc_var_quotient}
\title{Variance of the quotient of two variables}
\usage{
calc_var_quotient(x, y)
}
\arguments{
\item{x}{vector of values, of the same length of y}

\item{y}{vector of values, of the same length of x}
}
\value{
The variance of the quotient of the two variables (variable corresponding to first vector divided by variable
        corresponding to second vector)
}
\description{
Variance of the quotient of two variables
}
