% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guam_BB_expansion.R
\name{df_expansion}
\alias{df_expansion}
\title{Expansion and Species Composition of an expansion domain}
\usage{
df_expansion(
  year,
  pool_f,
  species,
  bl,
  bl2,
  bl_unknown_method,
  bl_unknown_fished,
  bl_known_fished,
  p1,
  p2,
  sample_days,
  days,
  interviews,
  interviews_raw,
  catch,
  catch_raw,
  methods_refer,
  ports_refer,
  reference
)
}
\arguments{
\item{year}{Year of the expansion}

\item{pool_f}{Whether to pool interviews when fewer than 3 are available for a domain}

\item{species}{a vector of species or all species (when spcies = NA)}

\item{bl}{boat log entries ...}

\item{bl2}{squared counts of boat log entries with known fishing method for each domain ([PORT_FK, TYPE_OF_DAY, METHOD_FK, CHARTER_F])}

\item{bl_unknown_method}{counts of boat log entries with unknown fishing method for each domain ([PORT_FK, TYPE_OF_DAY, CHARTER_F])}

\item{bl_unknown_fished}{counts of boat log entries with unknown fishing status for each domain ([PORT_FK, TYPE_OF_DAY, CHARTER_F])}

\item{bl_known_fished}{counts of boat log entries with known fishing status for each domain ([PORT_FK, TYPE_OF_DAY, CHARTER_F])}

\item{p1}{adjustment factors to account for missed vessels outside of shift times for each domain ([PORT_FK, METHOD_FK, CHARTER_F])}

\item{p2}{adjustment factor to account for vessels outside of the sampled ports}

\item{sample_days}{the number of sample days for each domain ([PORT_FK, TYPE_OF_DAY])}

\item{days}{the number of calendar days for each domain ([TYPE_OF_DAY])}

\item{interviews}{interviews for the expansion period}

\item{interviews_raw}{all interviews for all years, for use when pooling}

\item{catch}{interview catch data for the expansion period}

\item{catch_raw}{interview catch data for all years}

\item{methods_refer}{vector to convert method keys to method indices}

\item{ports_refer}{vector to convert port keys to port indices}

\item{reference}{reference values for use when insufficient interviews are available for a domain}
}
\value{
Two data frames containing all expansion and species composition fields for each expansion domain
}
\description{
Expansion and Species Composition of an expansion domain
}
