% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expalg.cie-package.R
\docType{package}
\name{expalg.cie-package}
\alias{expalg.cie-package}
\alias{_PACKAGE}
\alias{expalg-package}
\title{expalg.cie: US Territorial Expansion Algorithm}
\description{
This package contains the methods to run the Guam boat-based expansion algorithm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PIFSCstockassessments/expalg.cie}
  \item Report bugs at \url{https://github.com/PIFSCstockassessments/expalg.cie/issues}
}

}
\author{
\strong{Maintainer}: Eric Fletcher \email{Eric.Fletcher@noaa.gov}

Authors:
\itemize{
  \item Toby Mathews \email{toby.e.matthews@noaa.gov}
  \item Hongguang Ma \email{hongguang.ma@noaa.gov}
  \item Felipe Carvalho \email{felipe.carvalho@noaa.gov}
}

}
\keyword{internal}
