% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g_bb_spc}
\alias{g_bb_spc}
\alias{guamBB_species}
\alias{guamspecies}
\title{Guam Species Composition Dataset}
\format{
A data.frame with 72049 observations with 25 variables:
\describe{
\item{SPC_PK}{The private key associated with a particular species catch expansion.}
\item{EXP_FK}{The private key associated with a particular domain expansion.}
\item{FLAG_POOLED}{Flag to include pooled Interviews.}
\item{SPECIES_FK}{The foreign key associated with a particular species.}
\item{NUM_INTERVIEW}{The number of interviews of the current domain that included catch of the current species.}
\item{NUM_INTERVIEW_POOLED}{The number of boat log entries (excluding \code{METHOD_FK}) where the fishing method was unknown.}
\item{KGS_CAUGHT}{The estimated expanded catch for the current species \emph{s} in the current domain. }
\item{NUM_KEPT}{The number of specimens of the current species \emph{s} that were recorded in interviews of the current domain.}
\item{PRICE_LB}{Price per weight}
\item{SYS_CREATE_TIME}{System Time Entry was created.}
\item{SYS_MODIFY_TIME}{System Time Entry was modified.}
\item{HASH}{Hash signature.}
\item{CPUE_TRIP}{The average catch-per-trip from interviews for the current domain.}
\item{CPUE_HOUR}{The average catch-per-hour from interviews for the current domain. Measured in kg/hr.}
\item{CPUE_GEAR}{The average catch-per-gear from interviews for the current domain. Measured in kg/gear.}
\item{CPUE_FISHER_HOUR}{The average catch-per-fisher-hour from interviews for the current domain. Measured in kg/(fisher * hour).}
\item{VAR_CPUE_TRIP}{The variance of the catch-per-trip from interviews for the current domain. Measured in (kg/trip)^2. }
\item{VAR_KGS_CAUGHT}{The variance of the estimated expanded catch for the current species \emph{s} in the current domain.}
\item{STD_CPUE_TRIP}{The coefficient of variation of the catch-per-trip from interviews for the current domain. Measured in kg/trip.}
\item{VAR_CPUE_HOUR}{The variance of the catch-per-hour from interviews for the current domain. Measured in (kg/hr)^2.}
\item{STD_CPUE_HOUR}{The coefficient of variation of the catch-per-hour from interviews for the current domain. Measured in kg/hr.}
\item{VAR_CPUE_GEAR}{The variance of the catch-per-gear from interviews for the current domain. Measured in (kg/gear)^2.}
\item{STD_CPUE_HOUR.1}{The coefficient of variation of the catch-per-hour from interviews for the current domain. Measured in kg/hr.}
\item{VAR_CPUE_FISHER_HOUR}{The variance of the catch-per-fisher-hour from interviews for the current domain. Measured in (kg/(fisher * hour))^2.}
\item{STD_CPUE_FISHER_HOUR}{The coefficient of variation of the catch-per-fisher-hour from interviews for the current domain. Measured in kg/(fisher * hour).}
}
}
\usage{
g_bb_spc
}
\description{
Representing the total catch by species for each estimation domain. See the vignette \emph{"Guam Boat Based Expansion Database Fields"} for more information.
}
\keyword{datasets}
