% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guam_BB.R
\name{get_charter}
\alias{get_charter}
\title{Charter status index corresponding to a database charter entry}
\usage{
get_charter(charter)
}
\arguments{
\item{charter}{a single string or boolean for charter status}
}
\value{
Charter indices for the input string/boolean (1 = charter, 2 = non-charter)
}
\description{
Charter status index corresponding to a database charter entry
}
