% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guam_BB.R
\name{get_charters}
\alias{get_charters}
\title{Charter status indices corresponding to a vector of database charter entries}
\usage{
get_charters(charters)
}
\arguments{
\item{charters}{vector of strings/boolean or single string/boolean for charter status}
}
\value{
Charter indices for the input string(s)/boolean(s) (1 = charter, 2 = non-charter)
}
\description{
Charter status indices corresponding to a vector of database charter entries
}
