% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guam_BB.R
\name{get_days}
\alias{get_days}
\title{Type of day indices corresponding to a vector of database type of day entries}
\usage{
get_days(type_of_days)
}
\arguments{
\item{type_of_days}{vector of strings or single string for type of day ("WD" or "WE")}
}
\value{
The type of day indices for the input string(s) (1 = "WD", 2 = "WE")
}
\description{
Type of day indices corresponding to a vector of database type of day entries
}
