% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guam_BB_expansion.R
\name{get_survey_zone}
\alias{get_survey_zone}
\title{Survey Zone key corresponding to port}
\usage{
get_survey_zone(port_fk)
}
\arguments{
\item{port_fk}{Single integer for port key}
}
\value{
The survey zone key corresponding to the input port key
}
\description{
Survey Zone key corresponding to port
}
