% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guam_BB_interview_pooling.R
\name{pool_interviews}
\alias{pool_interviews}
\title{Perform interview pooling for a domain}
\usage{
pool_interviews(
  interviews,
  year,
  method,
  type_of_day,
  port,
  charter,
  unsampled_ports
)
}
\arguments{
\item{interviews}{All interviews for all years}

\item{year}{Year of the expansion}

\item{method}{Fishing method of the current stratum expansion}

\item{type_of_day}{Type of day of the current stratum expansion}

\item{port}{Port of the current stratum expansion}

\item{charter}{Charter status of the current stratum expansion}

\item{unsampled_ports}{Boolean representing whether the current pooling is for unsampled ports}
}
\value{
data frame containing the pooled interviews for the current stratum; if pooling could not produce at least 3 interviews, NA is returned
}
\description{
Function to conduct interview pooling. It is called by code in the file \code{"R/guam_ BB_ expansion.R"}.
}
