% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guam_BB.R
\name{run_expalg}
\alias{run_expalg}
\title{Run Expansion Algorithm}
\usage{
run_expalg(pool_f = TRUE, species = NA, start_year = 1982, end_year = 2019)
}
\arguments{
\item{pool_f}{defaults to T, which indicates that interview pooling is used; change to F if pooling should not be used}

\item{species}{NA to expand for all species, or a vector of SPECIES_PK to only include specific species}

\item{start_year}{the first year of the expansion to run (earliest available year is 1982)}

\item{end_year}{the last year of the expansion to run (most recent available year is 2019)}
}
\value{
List containing 2 data frames:
\describe{
\item{bb_exp}{Total catch for each estimation domain}
\item{bb_spc}{Total catch by species for each estimation domain}
}
}
\description{
Run Expansion Algorithm
}
\note{
Whether to pool interviews when fewer than 3 are available for a stratum
}
